+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Version 1.2.23
 
-  * 
+  * Geting list of user roles for getting sections and permissions now done using the get_user_roles method of api instances
+  * get_user_roles will raise an Osm::NoActiveRoles exception if the user has no active roles
 
 ## Version 1.2.22